# docs/swagger.yaml
definitions:
...
  CreateBraineryLogRequest:
    properties:
      discordID:
        type: string
      githubID:
        type: string
      publishedAt:
        type: string
      reward:
        type: number
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      url:
        type: string
    required:
    - discordID
    - publishedAt
    - reward
    - tags
    - title
    - url
    type: object
...
  ErrorResponse:
    properties:
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/view.ApiError'
        type: array
    type: object
...
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  view.ApiError:
    description: validation error details
    properties:
      enums:
        description: available options incase of field's payload is enums
        items:
          type: string
        type: array
      field:
        description: the field cause the error
        type: string
      msg:
        description: error message
        type: string
    type: object
...
info:
  contact:
    email: benjamin@d.foundation
    name: Nam Nguyen
    url: https://d.foundation
  description: This is api document for fortress project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FORTRESS API DOCUMENT
  version: v0.1.39
paths:
...
  /brainery-logs:
    post:
      consumes:
      - application/json
      description: Create brainery logs
      operationId: createBraineryLog
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CreateBraineryLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create brainery logs
      tags:
      - Project
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"